Add files via upload
Business Objective:
A US-based housing company named Surprise Housing has decided to enter the Australian market. The company is looking at prospective properties to buy to enter the market.

The company wants to know:
1. Which variables are significant in predicting the price of a house
2. How well those variables describe the price of a house
3. Based on the important driving factors the company wants to manipulate the strategy of the firm and concentrate on areas that will yield high returns.

Analytical Objective:
1. Build a regression model using regularisation in order to predict the actual value of the prospective properties and decide whether to invest in them or not.
2. Determine the optimal value of lambda for ridge and lasso regression.
3. Identify which variables are significant in predicting the price of a house.
4. Understand how well those variables describe the price of a house.

Further, the model will be a good way for management to understand the pricing dynamics of a new market.